Pinned the version
This is required for Maven Central releases. Otherwise a dynamic version ends up in pom and Central does not like it. There is no simple solution to put "resolved" versions in the pom (I found this: https://stackoverflow.com/questions/20959558/in-gradle-how-can-i-generate-a-pom-file-with-dynamic-dependencies-resolved-to-t). Instead of complicating the build logic, let's just pin the version for now.

Fixes part of #3